requirements.txt
Flask==2.3.3
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.32.5
Pillow==11.3.0
